Defender XDR Audit/xdr_requirements.txt
azure-identity>=1.12.0
azure-mgmt-subscription>=3.1.1
azure-core>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0